+++ backend/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feature: use Hitobito course state to determine if application is open for courses
 - feature: add quick filters for course categories
 
 ## Version 1.0.15, 23.03.2026